Add acceptance tests for puppet-diskimage_builder.
Add acceptance tests for puppet-diskimage_builder module so that once
the module is applied we check if packages were installed.

During the testing, we noticed that apt::ppa was failing for two
reasons, it was being used without including the module and it was not
updating the package list after including the ppa.

We then included the apt module in the beginning and required an apt
update before installing the packages.

Change-Id: I37a7e4ae9f1a380568c0aa4f09cef1dfb7c4aaa9
Co-Authored-By: Bruno Tavares <btavare@thoughtworks.com>