Add gnupg2 install for zypper as well
As of today zypper is explicitly using /usr/bin/gpg2 for signature
validation but this is typically not installed on an Ubuntu host
because it defaults to gnupg (1.x, which only provides /usr/bin/gpg).

This change is needed so that the zypper-minimal element has a
chance of working on a Xenial host.

Change-Id: If57b07dcf31b8eeb2a3a44442315dd1d70b3523d